changed all roastId to long...in MANY files. all tests pass (except getMostRecient(), which was failing already

+++ app/src/androidTest/java/com/andrewkjacobson/android/roastassistant/db/dao/ReadingDaoTest.java
@@ -51,7 +51,7 @@ public class ReadingDaoTest extends TestCase {
 
     @Test
     public void testGet() {
-        int roastId = expected.get(0).getRoastId();
+        long roastId = expected.get(0).getRoastId();
         int seconds = expected.get(0).getSeconds();
         LiveData<ReadingEntity> retrieved = readingDao.get(roastId, seconds);
