+++ app/src/main/java/com/andrewkjacobson/android/roastassistant/db/dao/CrackReadingDao.java
@@ -9,8 +9,8 @@ import java.util.List;
 @Dao
 public abstract class CrackReadingDao extends BaseDao<CrackReadingEntity> {
         @Query("SELECT * FROM crack_reading_entity WHERE roastId=:roastId ORDER BY seconds ASC")
-        public abstract LiveData<List<CrackReadingEntity>> getAll(int roastId);
+        public abstract LiveData<List<CrackReadingEntity>> getAll(long roastId);
 
         @Query("SELECT * FROM crack_reading_entity WHERE roastId=:roastId AND seconds=:seconds")
-        public abstract LiveData<CrackReadingEntity> get(int roastId, int seconds);
+        public abstract LiveData<CrackReadingEntity> get(long roastId, int seconds);
 }
