+++ app/src/main/java/com/andrewkjacobson/android/roastassistant/db/dao/ReadingDao.java
@@ -10,13 +10,13 @@ import java.util.List;
 @Dao
 public abstract class ReadingDao extends BaseDao<ReadingEntity> {
     @Query("SELECT * FROM reading_entity WHERE roastId=:roastId ORDER BY seconds ASC")
-    public abstract LiveData<List<ReadingEntity>> getAll(int roastId);
+    public abstract LiveData<List<ReadingEntity>> getAll(long roastId);
 
     @Query("SELECT * FROM reading_entity WHERE roastId=:roastId AND seconds=:seconds")
-    public abstract LiveData<ReadingEntity> get(int roastId, int seconds);
+    public abstract LiveData<ReadingEntity> get(long roastId, int seconds);
 
     @Query("DELETE FROM reading_entity WHERE roastId=:roastId AND seconds=:seconds")
-    public abstract void delete(int roastId, int seconds);
+    public abstract void delete(long roastId, int seconds);
 
     @Query("DELETE FROM reading_entity")
     public abstract void deleteAll();
