+++ app/src/main/java/com/andrewkjacobson/android/roastassistant/model/Roast.java
@@ -1,8 +1,8 @@
 package com.andrewkjacobson.android.roastassistant.model;
 
 public interface Roast {
-    int getId();
-    void setId(int id);
+    long getId();
+    void setId(long id);
     boolean isRunning();
     void setRunning(boolean isRunning);
     boolean isFinished();
